Fixes handling of methods that contain a comment on method definition line (#17)
* Fixes handling of methods that contain a comment on method definition
line

Before when we saw a method that contained a comment on the same line as
its method definition the method was deemed unprocesseable and was
skipped. As a result any signatures or documentation on the method would
be replaced by the comment. This commit removes an extraneous check that
deems the method processable.

* more tests & changelog

* Test command statements

Co-authored-by: Douglas Eichelberger <dduugg@gmail.com>